fix(currency): change the type to be compatible with any other database

+++ app/Database/Migrations/2023-07-22-134147_CreateCurrencies.php
@@ -15,7 +15,7 @@ class CreateCurrencies extends Migration
                 "auto_increment" => true,
             ],
             "user_id" => [
-                "type" => "BIGINT",
+                "type" => "INT",
                 "unsigned" => true,
             ],
             "code" => [
